LayoutTests/http/tests/site-isolation/domwindow-length-with-remote-frames.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script>
    if (window.testRunner) {
        testRunner.waitUntilDone();
        testRunner.dumpAsText();
    }

    function loadNextFrame() {
        const frames = [
            'http://127.0.0.1:8000/site-isolation/resources/domwindow-length-tester.html?frame=same1',
            'http://localhost:8000/site-isolation/resources/domwindow-length-tester.html?frame=cross1',
            'http://127.0.0.1:8000/site-isolation/resources/domwindow-length-tester.html?frame=same2',
            'http://localhost:8000/site-isolation/resources/domwindow-length-tester.html?frame=cross2'
        ];

        let currentFrame = 0;
        (function createFrame() {
            if (currentFrame >= frames.length) {
                if (window.testRunner) { testRunner.notifyDone(); }
                return;
            }

            const iframe = document.createElement('iframe');
            iframe.src = frames[currentFrame];
            iframe.onload = () => {
                currentFrame++;
                setTimeout(createFrame, 100);
            };
            document.body.appendChild(iframe);
        })();
    }

    window.onload = loadNextFrame;
</script>
